Updated the readme, literally replaced "not implemented" with "in development". Lol.
inputValidator seems to work now, EXCEPT if you don't input the dice type at all i.e. "1d+3".

Need to think if it's possible to search by array .split function?

if array size x then incorrect input.

After that, might as well merge to beta.

Been thinking of creating separate print and report methods for the rolls so if the user wants to see only the final roll it is possible. And vice versa  if  the user want's to see every roll, it's possible.