lol noticed how terrible the code was with ArrayList so I replaced it with normal Array.
inputSpammer still Work In Progress but I think I got the normall Array replaced very neatly.
Some parts of the code actually feel easier to read, just needed to paste the normal Array report syntax logic so you don't have to scroll all the way to the main to see it.

Commit and push for now. Let's see if this can get to the main branch eventually.