Updated the launch files, now users can actually use them to get to the program.
The users are able to get tutorial's 2nd info screen.
After proceeding that screen the program will crash.

Bug noticed:
when entering to the tutorial and trying to go back by giving the input 'X'
the program doesn't seem to go back to the main menu.
Or atleast print the menu, worth to checkout if there's a while loop going around!